requirements.txt
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
scikit-learn==1.5.0
presidio-anonymizer==2.2.351
presidio-analyzer==2.2.351
faker==19.6.2
ipython==8.15.0
textdistance==4.6.0
gensim==4.3.2
scipy==1.11.2
transformers==4.33.2
langchain==0.0.335
python-dotenv==1.0.0
python-docx==1.1.2
pyzmq==25.1.2
nltk==3.8.1
sentence-transformers==2.2.2
huggingface_hub==0.15.1